When we’re denormalizing relationships onto entities, and we encounter a circular reference, we should leave the leaf as a valid JSON-API object (with type and id) rather than undefined.
Also, we should avoid barfing if passed an invalid ref, and instead just return `undefined` (this is especially important when receiving a JSON-API response that has a ref in a relationship, but that object is not included in the `included` section. Cartographer never does this, but it is allowed by the spec (and is even featued in the jsonapi.org front-page example)).