Localize DenormalizationCache inside of store.
Nion was not explicitly designed to run on the server. The clearest
indication of this is the `DenormalizationCache` singleton, which uses
an unbounded set of objects to hang onto denormalized entities.

This works great on the client, but leaks memory on the server unless
explicitly purged. With the upcoming adoption of React Server
Components, we no longer have a lifecycle hook to flush the cache as
rendering can happen concurrently across trees.

The safest thing to do here is to localize the cache to the store
itself. Thankfully there is a single entrypoint at which we access this
cache – a selector – which we can take advantage of for our use case.